graceful shutdown on Linux and macOS
Calling `loop.stop()` from a Pyro5 callback works on macOS with CFRunLoop integration but not on linux. Calling `loop.call_soon_threadsafe(loop.stop)` works on Linux but not with CFRunLoop integration on macOS. We therefore use a future `shutdown_complete` to request a shutdown and run the loop until `shutdown_complete` has a result.