Resolves locale issue on OpenBSD
OpenBSD only supports one value of LC_CTYPE: C.  For applications which really want to use UTF-8 it also supports en_US.UTF-8 as an acceptable value (and implicitly converts anything ending with .UTF-8 to en_US.UTF-8). See OpenBSD's man 1 locale.

When spawning the daemon maestral execs it in a custom environment with the value of LC_CTYPE set UTF-8 which isn't valid on this platform so the daemon fallsback to running in ASCII mode... which breaks everything.

Unfortunately the call to env.update in daemon.py still seems to break things even if the LC_CTYPE is set correctly, so for OpenBSD I've disabled it.  Things seem to run fine otherwise.